auth_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
PyJWT==2.8.0
python-multipart==0.0.6
pydantic[email]==2.5.0
bcrypt==4.1.2
